Sync aclui, advapi32, atl, authz, kernel32, msi, oledlg, powrprof, qmgr, riched20, samlib to Wine 1.2rc5 Update some psdk Headers to get some more synched winetests build
svn path=/trunk/; revision=47930

+++ reactos/dll/win32/riched20/rtf.h
@@ -1122,7 +1122,7 @@ struct _RTF_Info {
     char *rtfTextBuf;
     int	rtfTextLen;
 
-    long rtfLineNum;
+    int rtfLineNum;
     int	rtfLinePos;
